+++ bigmler/defaults.py
@@ -106,7 +106,10 @@ FLAGS = {'BigMLer': [{'flag': 'debug', 'type': 'boolean'},
                      {'flag': 'test_types', 'type': 'string'},
                      {'flag': 'test_source', 'type': 'string'},
                      {'flag': 'test_dataset', 'type': 'string'},
-                     {'flag': 'no_batch', 'type': 'boolean'}]}
+                     {'flag': 'no_batch', 'type': 'boolean'},
+                     {'flag': 'dataset_attributes', 'type': 'string'},
+                     {'flag': 'output', 'type': 'string'},
+                     {'flag': 'new_fields', 'type': 'string'}]}
 
 
 def get_user_defaults(defaults_file=DEFAULTS_FILE):