+++ bigmler/processing/models.py
@@ -34,10 +34,8 @@ from bigml.fields import Fields
 from bigmler.processing.ensembles import (ensemble_processing,
                                           ensemble_per_label)
 
+from bigmler.processing.ensembles import MISSING_TOKENS
 
-MISSING_TOKENS = ['', 'N/A', 'n/a', 'NULL', 'null', '-', '#DIV/0', '#REF!',
-                  '#NAME?', 'NIL', 'nil', 'NA', 'na', '#VALUE!', '#NULL!',
-                  'NaN', '#N/A', '#NUM!', '?']
 MONTECARLO_FACTOR = 200
 
 
@@ -152,7 +150,7 @@ def models_processing(datasets, models, model_ids, objective_field, fields,
             # Ensemble of models
             (ensembles, ensemble_ids,
              models, model_ids, resume) = ensemble_processing(
-                 dataset, objective_field, fields, api, args, resume,
+                 datasets, objective_field, fields, api, args, resume,
                  name=name, description=description, model_fields=model_fields,
                  session_file=session_file, path=path, log=log)
             ensemble = ensembles[0]
@@ -161,7 +159,7 @@ def models_processing(datasets, models, model_ids, objective_field, fields,
 
         else:
             # Set of partial datasets created setting args.max_categories
-            if len(datasets) > 1:
+            if len(datasets) > 1 and args.max_categories:
                 args.number_of_models = len(datasets)
             # Cross-validation case: we create 2 * n models to be validated
             # holding out an n% of data